게시글 삭제 기능 구현
Article 클래스에 deleted 필드를 넣어 게시글 삭제 시 deleted가 false가 되도록 구현
게시글에 댓글이 존재하면 삭제할 수 없음

+++ src/main/java/codesquad/springcafe/service/ArticleService.java
@@ -2,21 +2,26 @@ package codesquad.springcafe.service;
 
 import codesquad.springcafe.dto.ArticleRequestDto;
 import codesquad.springcafe.exception.ArticleNotFountException;
+import codesquad.springcafe.exception.CannotDeleteArticleWithRepliesException;
 import codesquad.springcafe.model.Article;
+import codesquad.springcafe.model.Reply;
 import codesquad.springcafe.repository.ArticleRepository;
+import codesquad.springcafe.repository.ReplyRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final ReplyRepository replyRepository;
 
-    public ArticleService(ArticleRepository articleRepository) {
+    public ArticleService(ArticleRepository articleRepository, ReplyRepository replyRepository) {
         this.articleRepository = articleRepository;
+        this.replyRepository = replyRepository;
     }
-
 
     public void createArticle(Article article) {
         articleRepository.save(article);
@@ -40,6 +45,10 @@ public class ArticleService {
     }
 
     public void delete(Long articleId) {
+        List<Reply> replies = replyRepository.findRepliesByArticleId(articleId);
+        if (!replies.isEmpty()) {
+            throw new CannotDeleteArticleWithRepliesException();
+        }
         articleRepository.delete(articleId);
     }
 }
